+++ test/blackbox-tests/test-cases/github8041.t
@@ -17,12 +17,6 @@
 If sendfile fails, we should fallback to the portable implementation.
 
   $ strace -e inject=sendfile:error=EINVAL -o /dev/null dune build @install
-  Error: _build/default/data2.txt: Permission denied
-  -> required by _build/default/data2.txt
-  -> required by _build/install/default/share/p/data2.txt
-  -> required by _build/default/p.install
-  -> required by alias install
-  [1]
 
 #8210: data2.txt is copied from readonly-file data.txt (#3092), so it needs to
 be adequately unlinked before starting the fallback.